Refactor search bar to header
Refactors the search bar component from the hero section and implements it in the header. The header search will be an icon that expands into an input field upon clicking, and will be visible on pages other than the homepage and user profile.
